JuliaOptimizationChallenge

Documentation for JuliaOptimizationChallenge.

JuliaOptimizationChallenge.Problem

JuliaOptimizationChallenge.Problem

JuliaOptimizationChallenge.compute_upper_bound

JuliaOptimizationChallenge.find_lookforward_paths_parallel

JuliaOptimizationChallenge.solver_parallel_lookforward

JuliaOptimizationChallenge.Problem — Type

Store a problem which represents a city using a matrix and other parameters

Fields

- total_duration::Int: total time allotted for the car itineraries (in seconds)
- nb_cars::Int: number of cars in the fleet
- starting_junction::Int: junction at which all the cars are located initially
- junctions::Vector{Junction}: list of junctions
- adjacency::Vector{Vector{Street}}: list where ith entry is all the streets that you can travel through from ith junction

JuliaOptimizationChallenge.Problem — Method

Creates a Problem instance using a HashCode2014.jl City object

JuliaOptimizationChallenge.compute_upper_bound — Method

Calculate an upper bound on the number of meters that can be covered using a HashCode2014.jl City instance

JuliaOptimizationChallenge.find_lookforward_paths_parallel — Method

Helper function that finds all possible paths at certain junction, with the current visited street.

Output

- the optimal path at N steps

JuliaOptimizationChallenge.solver_parallel_lookforward — Method

Find a possible solution of itineraries for each car, given a problem instance

Output

- a set of itineraries for each of the cars